+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @opencx/widget-react
 
+## 4.0.56
+
+### Patch Changes
+
+- add onClicked option for header buttons
+- Updated dependencies
+  - @opencx/widget-core@4.0.56
+  - @opencx/widget-react-headless@4.0.56
+
 ## 4.0.55
 
 ### Patch Changes